refactoring: auth domain dependency set
