+++ src/domains/mod.rs
@@ -1,2 +1,3 @@
-pub mod commands;
+pub mod auth;
 pub mod models;
+pub mod post;
